requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-cloud-pubsub==2.21.5
python-dotenv==1.0.1
google-cloud-discoveryengine==0.12.3
